+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ## 1.8 branch
 ### 1.8.2
+* `TestTrait::assertException()` is deprecated and will be removed in a later release;
 * `TestTrait::expectAssertionFailed()` is deprecated and will be removed in a later release.
 
 ### 1.8.1