+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## 1.8 branch
 ### 1.8.2
+* added tests for PHP 8.2;
 * `BodyParser` is deprecated and will be removed in a later release;
 * `TestTrait::assertException()` is deprecated and will be removed in a later release;
 * `TestTrait::expectAssertionFailed()` is deprecated and will be removed in a later release.